stat works without time period
